Fix default of 'closed' to stop writing open paths as implicitly closed
As discussed in schriftgestalt/GlyphsSDK#92, Glyphs asssumes that the
'closed' property value is true when omitted not false, and so this
commit changes our default value to ensure that this is respected when
reading and writing path objects.

In addition, Glyphs _always_ serialises the property, and so this commit
enforces that also to ensure that the closedness of a path is always
explicit.